Add blog scripts and a button with click handlers
We tested out scripts for the index of the blogs and created the button
and event handlers. we also made a blogs folder for api events and ui
and set up the index actions for them.

Received a successful response from the server, but the data is not
rendering on the page yet.

Contributors: J.A.S

Next Steps: Get the data to render on the page through the handlebars