Get blog index to render in handlebars
When the index button is clicked, the response from the server gets
rendered into handlebars.  The event function checks for if a user is
logged in.  If an account is logged in, the posts that they created
will be displayed.  If no user is logged in, all blog posts will be
displayed.  This is so guests can see content prior to logging in.

Next Steps:  Fix date/timestamp in backend models.
Make create form on front end for blogs.

Contributors: J.A.M.S.