Fix date render in handlebars blogs
Fixed the createdAt date from displaying in handlebars as a full
timestamp to only displaying the year month and day it was created at.
This should reduce confusion for the user.  The ui event processes the
response and reassigns the new date to the data that gets sent to
handlebars.

Next Steps: Create forms for blog

Contributors:  J.A.M.S.