Add show functionality for pages
Index shows the title of the page.  When the title is clicked, the
show event runs, calling the api show, and the response is rendered
in handlebars.  So when a title is clicked, the content of that page
replaces the list of titles.

Next: Update page

Contributors: J.A.M.S.