Format tabs and add all-blog handlebars
We grouped the content into home, blog, and page tabs.  blog and page
tabs are only visible while signed in.  For guests, we made a new
handlebars and div to show all user's blogs, but we removed the
update and delete functionalities from this handlebars.

Next step: clear forms

Contributors: J.A.M.S.